+++ pkgs/ffigen/lib/src/config_provider/yaml_config.dart
@@ -1270,15 +1270,6 @@ final class YamlConfig {
       importType: importType,
       objectiveC: language == Language.objc
           ? ObjectiveC(
-              interfaces: Interfaces(
-                includeTransitive: includeTransitiveObjCInterfaces,
-              ),
-              protocols: Protocols(
-                includeTransitive: includeTransitiveObjCProtocols,
-              ),
-              categories: Categories(
-                includeTransitive: includeTransitiveObjCCategories,
-              ),
               externalVersions: externalVersions,
               // ignore: deprecated_member_use_from_same_package
               generateForPackageObjectiveC: generateForPackageObjectiveC,
@@ -1390,6 +1381,7 @@ final class YamlConfigAstVisitor extends public_ast.Visitor {
     if (config.interfaceModule(_decl(node)) case final module?) {
       node.module = module;
     }
+    node.includeCategories = config.includeTransitiveObjCCategories;
   }
 
   @override