+++ pkgs/ffigen/lib/src/visitor/list_bindings.dart
@@ -65,14 +65,9 @@ class ListBindingsVisitation extends Visitation {
 
   @override
   void visitObjCInterface(ObjCInterface node) {
-    final _IncludeBehavior includeBehavior;
-    if (node.isInternal) {
-      includeBehavior = _IncludeBehavior.transitive;
-    } else if (config.objectiveC?.interfaces.includeTransitive ?? false) {
-      includeBehavior = _IncludeBehavior.configOrTransitive;
-    } else {
-      includeBehavior = _IncludeBehavior.configOnly;
-    }
+    final includeBehavior = node.isInternal
+        ? _IncludeBehavior.transitive
+        : _IncludeBehavior.configOnly;
     final omit = node.unavailable || !_visitImpl(node, includeBehavior);
 
     if (omit && directTransitives.contains(node)) {
@@ -94,7 +89,7 @@ class ListBindingsVisitation extends Visitation {
   @override
   void visitObjCCategory(ObjCCategory node) => _visitImpl(
     node,
-    config.objectiveC?.categories.includeTransitive ?? false
+    node.parent.includeCategories
         ? _IncludeBehavior.configOrDirectTransitive
         : _IncludeBehavior.configOnly,
   );
@@ -102,13 +97,7 @@ class ListBindingsVisitation extends Visitation {
   @override
   void visitObjCProtocol(ObjCProtocol node) {
     final omit =
-        node.unavailable ||
-        !_visitImpl(
-          node,
-          config.objectiveC?.protocols.includeTransitive ?? false
-              ? _IncludeBehavior.configOrTransitive
-              : _IncludeBehavior.configOnly,
-        );
+        node.unavailable || !_visitImpl(node, _IncludeBehavior.configOnly);
 
     if (omit && directTransitives.contains(node)) {
       node.generateAsStub = true;